Split `util.serde` into submodules Move `ModelCheckpointType` and co to `util.serde.checkpoint`
